backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
thefuzz>=0.22.0
python-dateutil>=2.9.0
